SDDL-104: Add comparison function

+++ app.py
@@ -3,20 +3,34 @@ import urllib.request
 from bs4 import BeautifulSoup
 from collections import OrderedDict
 import json
+from os import path
+import os.path
+import sys
 
 DragonRouteUrl = "https://drexel.edu/facilities/transportation/busServiceSchedules/DragonRoute/"
 PoweltonSpringGardenRouteUrl = "https://drexel.edu/facilities/transportation/busServiceSchedules/PoweltonSpringGardenRoute/"
 QueenLaneRouteUrl = "https://drexel.edu/facilities/transportation/busServiceSchedules/QueenLaneRoute/"
 
-def pushToDb(data):
+def pushToDb(data, name):
     print()
 
-def comparison(data):
-    print()
+def comparison(data, name):
+    #open the file for compare
+    original = open(os.path.join(sys.path[0], name + '.json'), 'r')
+    originalJsonObj = json.loads(original.read())
+    compareAgainstJsonObj = json.loads(data)
+
+    a, b = json.dumps(originalJsonObj), json.dumps(compareAgainstJsonObj)
+    if not a == b:
+        pushToDb(data, name)
 
 def constructShape(data, name):
-    with open(name + '.json','w') as outfile:
-        json.dump(data, outfile)
+    if not path.exists(name + '.json'):
+        with open(name + '.json','w') as outfile:
+            json.dump(data, outfile)
+        pushToDb(data)
+    else:
+        comparison(json.dumps(data), name)
 
 def getData(url, name):
     response = requests.get(url)
@@ -42,6 +56,6 @@ def getData(url, name):
     constructShape(data, name)
 if __name__ == "__main__":
     getData(DragonRouteUrl, "DragonRoute")
-    #getData(PoweltonSpringGardenRouteUrl)
-    #getData(QueenLaneRouteUrl)
+    #getData(PoweltonSpringGardenRouteUrl, "PoweltonSpringGardenRoute")
+    #getData(QueenLaneRouteUrl, "QueenLaneRoute")
